fix(compaction): break manager->runner->tools runtime cycle failing CI
check-runtime-cycles counts dynamic import() as a runtime edge, so
manager.ts's lazy imports of compaction helpers from subagent-runner.ts
closed the loop manager -> subagent-runner -> tools/index -> manager
and failed the lint-and-test job on PR #141.

Extract the mid-run compaction helpers (buildSubagentPartialReport,
resolveSubagentContextTokens, buildSelectiveSubagentApply,
tryCompactSubagentHistory) into a dedicated subagent-compaction.ts
module with no tools/ or manager dependency, and import them statically
from the manager. subagent-runner.ts returns to being purely the stream
driver.
